+++ voxblox/include/voxblox/utils/protobuf_utils.h
@@ -2,7 +2,6 @@
 #define VOXBLOX_UTILS_PROTOBUF_UTILS_H_
 
 #include <fstream>
-
 #include <glog/logging.h>
 #include <google/protobuf/message.h>
 #include <google/protobuf/message_lite.h>
@@ -11,14 +10,14 @@ namespace voxblox {
 
 namespace utils {
 bool readProtoMsgCountToStream(std::fstream* stream_in, uint32_t* message_count,
-                               uint64_t* byte_offset);
+                               uint32_t* byte_offset);
 
 bool writeProtoMsgCountToStream(uint32_t message_count,
                                 std::fstream* stream_out);
 
 bool readProtoMsgFromStream(std::fstream* stream_in,
                             google::protobuf::Message* message,
-                            uint64_t* byte_offset);
+                            uint32_t* byte_offset);
 
 bool writeProtoMsgToStream(const google::protobuf::Message& message,
                            std::fstream* stream_out);
